refactor(e2e-test-kit): file-system helpers
